Update: Add Gantt class
The Gantt class represents the Gantt chart. It uses an int array to
model the chart. It implements methods to fill its model with process
information.
Its constructor receives Processes from the Parser, creates
a model based on total number of bursts and number of processes.
Its generate method receives an Algorithm instance, and fills the model,
which afterwards is a proper Gantt chart, albeit in a non-printable
form.