Upgrade: Completely rewrite Dispatcher::run
This commit if applied will delete vast swathes of code. Hopefully
most of it was redundant and this is an almost definitive modification.
Anyways, the class has lost like half its size, and presumably still
does the same as before. At least it does with FCFS.
The logic for updating the BCPs and the one for creating Gantt entries
have been properly isolated in their own separate methods, instead of
being smeared all over the run loop.
There is still testing to be done, specially with the other algorithms,
that will have to be modified because previously they had to take into
account that they would be called twice over the same process during
the loop setup.
That's no longer an issue. Agus you can finally stop whining about it.
And start doing it about the new issues this rework is sure to have
brought.