Feature: Add the CPU class
The CPU class is in charge of isolating all paging logic from
dispatcher. It receives all new instances related to paging and creates
entries for the 'Gantt' chart. The method usePage() simulates trying to
read from a page in RAM, and the following page fault that should occur
if said page weren't there.

The CPU class is bolted haphazardly to our program, and it's not really
well integrated into the hierarchy. If we were to do it right, it should
have a Dispatcher as one of its attributes, but that would entail a
refactor that we can't really afford now with the deadline less than
24hs away. It was easier this way, and it will work, so there's that.